Oasis expansion and cleanup
Oasis is a town that has been lived in with a large amount of people, and this now reflects that. Bathrooms have been added to houses and the Bar, the bar now has rooms they can rent out to wasters and more. Store lobby has been expanded and the house next to the store designed. Church and Oasis Oak has gone under total redesign, Poppy is of course added back. Underground section of Oasis looks less like a hobo den and more like an extension of the town, workshop added to forge area, lighting and flooring added all around, and one new room added for citizens.