Captains, HTs, and Logistics officer and more!
I have re-added the Captain and Heavy Troopers for the NCR, adjusted their loadouts, and in addition added the Logistics Officer as the Engineer counterpart of the Medical Officer. Gave the LT a R82 as well to reflect their higher rank, rather than using a service rifle. Gave LO and MO a custom 1911 to make them stand out more.