YuMail Express Ruin
Replacing the stop-gap ruin that was placed to fill in the gap from the followers outpost. Say hello to the YuMail Express, where the mailmen were planning to shoot apart the eyebots that were replacing them on the same day. They didn't have the chance to get embarrassed though, because the bombs dropped shortly beforehand.