Adds wires and signaler support to Universal Recorders (#467)
After seeing a reference to an unused wire panel in the description of the Universal Recorder, I decided to go ahead and add wires for the record, stop, and play verb that the taperecorder has. This will allow taperecorders to be controlled via remote signalling devices once attached.

Universal Recorders have always been lacking in my eyes, with hardly anyone (read - no one) using them. With this change, the taperecorders can now be used to construct a varied amount of neat signaller systems, finally giving them a purpose beyond news interviews. Here's a couple of ideas:

    An automated greeting system for stores, houses, or other places via a proximity sensor/infrared emitter and signaller assembly connected to the play button of the recorder
    A warning system that not only tells you when a proximity sensor has been triggered, but by setting a recording, where it has been triggered. (E.g. "Recording playing - "Enemy in the mines!")
    A message recording device for when you are away from home. Set one button on the wall to be the record signaller, and another to be stop, then leave the taperecorder where it can hear but not be touched by visitors.