Removing flask, Rebalancing Mutant Village (#473)
Removing the flask from wastelander spawn and putting the money purse where it would spawn in the pocket, a little easier for wastelanders who have to shift around all their belongings at the start of the round and removing the flask since its just clutter. Also moved the super mutants back to their original numbers in the village I made because it is meant to be twelve mutants, twelve beds. Spamming vault mutants and melee mutants around the place has no sense to it, so I undid that.