Apron, Cloak, Mantle rework, additions, restoration (#480)
Offshoot of the big armor/suits rework. Have to partition it or go insane.

What is this?
Basically expands the clothing layer options by adding the categories mantles and aprons to the accessory slot.
Both have a mainly aesthetic function, with some minor practical stuff like limited storage, small resist or climate bonuses.

Currently there is clothing called cloaks, mantles, capes that either go on suit or on accessory slot, either have togglable armored hood or not, either have thick armor or none, with zero visual or name difference.
This changes that issue, and adds both functionality and customization options.
If it has a hood and armor values - it is a cloak, goes on suit slot.
If it has no hood (nothing that cover the face/gives armor specifically) and no armor, its a mantle.

    Two new data files for these clothing added, two new icon files to contain the sprites as well.
    All clothing in current assets that fits into the categories and has acceptable sprites copied there and made to adhere to the templates, and all instances of cloak replaced with mantle for those items.
    Added a number of various storage options with special pockets for the differing types of clothing created.
    Current clothing types : Mantles, Ponchos, Aprons, Pelts (mantle style ones, not the head variants yet.)
    New art made for : Labor apron, Bartender apron, Housewife apron, Surgical apron, Poncho
    New clothing created + art : Mountain bear pelt, Fur pelt (longhorner), Ragged pelt (art the old unused lizard mantle so not mine), Gecko mantle.

Follow up for this PR will be some crafting recipes, replacing the items needed, and removing now obsolete items and art.