Loot tables added, spare magazines empty (#471)
Added a whole section of spawners with sub-lists themed on areas they are supposed to be placed in. Not actually mapped in, the amount of testing needed is more than I can do alone, having them in the repo allows other to test them and gradual implementation. It's modular in the same way as the old, just follows a different design philosophy and is much more up to date than the old lists, emphasis on the themes. Currently in alpha.

Most spare magazines that come with found guns on the map are now empty instead of filled. Don't affect cells or caseless ammo and a couple others. This affect the live map, part of reducing the inflation of ammo and guns, and it makes more sense the previous owner left it on the ground after having spent the ammo than being a nice fellow and loading it up before leaving it for a lucky waster. Gameplay wise it reduces autonomy a little bit, but with the amounts of ammo easily obtainable the effects should be modest, people will use backup guns a little more, and maybe feel a little more incentive to buy or find some ammo at some point.

Adjusted the money spawner percentages so its shifted more towards a bit smaller piles.