NCR Dungeon + Followers Work (#209)
Given the apparent added difficulty of the NCR pitstop dungeon, I've lowered the healths of the Ghouls by a smidgeon! I'm told the glowing ones have had less Nuka-Cola quantum intake due to the shortages...

Moreover, a few out-of-place mobs were removed, being Robobrains, Securitrons and Assaultrons. Neither of the three particularly belong in the Sonora, nevertheless TWO entities from the East Coast. What lore explains that, Desert Rose?

Lastly, the Followers don't particularly need to be mining at all. Their mining section has been reduced to an additional storage area, including a removal of the lining dense rock in the underground section to reduce metagaming and increase possibility. Metal and glass available to the Followers have been reduced to increase trade and communication with Oasis and or external sources. I'll see how this bodes over and consider reverting this!