Ghost Spawners (#391)
* Ghost Spawners

Adds two sets of ghost role spawners for the NCR, Legion and the BoS each. There's also a raider spawner.

* Actual names

Makes it so that if you spawn as a ghost role, your name isn't just the generic mob name (Eg. NCR Trooper). The proc is defined on specific paths, as it wouldn't make latin names for the legion, and also not to fuck up future ghost roles.