Adjusting Raider Ghost Role (#399)
* Adjust Raider Ghost Role

I'm seeing the value in adding the raider ghost role as a sort of antag to spice things up. I must see how players interact with this role to know how to move forward with it.

* Notifications

Notifies ghosts that the spawners are available.

* Small CoR adjustment

Refactored some of the values on the Cult of Rain mobs. They shouldn't be aaasss cancer to deal with, but still difficult.

* Sneaky commit

Adds a couple traits to the halloween species I added the other day. Don't want to make a whole new PR for this, so.

* Zombie nerf!

Whoops!

* Temporarily disables handheld radio crafting

For event

* Revert

* Mag change