BoS Loadout Update/New Knight Weapon (#436)
* BoS Loadout Update/New Knight Weapon

Added sidearms to Senior Paladins and Knights because they were some of the few combat roles not to have one.

Got rid of the Commando Carbine for Defender loadouts and replaced it with the new R93 PDW.

Added R93 loadout to Senior Knights as well under "Recon."

* Update automatic.dm

Changed the in-hand sprite of the Gauss Rifle to be more fitting than a shotgun.

* Update automatic.dm

* Update automatic.dm

Re-did inhand state for M72 to be more form-fitting.

* Update automatic.dm

Adding new line for linters.

* Update automatic.dm

Don't know how I missed this. Would have been OP as hell.

* Update automatic.dm

Did more playtesting, 9 round bursts w/ attachments isn't okay.

* Replaced BoS SWAT Masks with the security gas masks

Most Brotherhood players wanted the switch due to the security gas mask being able to be moved out of the way to actually eat. Plus it looks better.

* Update automatic.dm

Description change.

* Lame Chest Rigs replaced with chad assault belts

Chest rigs blocked nice looking icons. They were always thrown on the rack anyway for a belt, so this just streamlines that process.