Fixes the song editor UI on pianos/space minimoogs. (#64545)
This fixes a bug (#64516) where the song editor UI failed to update properly on pianos/space minimoogs.

As it turns out, the updateDialog call at the end of the file had nothing set for it, resulting in it runtiming any time it was supposed to be updating.

Setting it to (usr) fixed the problem, and the UI now works properly.

(cherry picked from commit 4d6eb42419b8d8c841671472d2604e65a3e763cc)