Adds in a deep water spot, to put on rivers and ponds to bucket up water.
Makes it so you don't need to have a sink in places where it doesn't make sense.

Puts some on the map, in key locations. One in the Khan farm, a few by the tent swamp area, and one in the legion camp.

Also removes the density on the smart fridge in the legion camp, so you can walk up to it. As well as removes the density on one of the pixal shifted signs in bighorn. Cause aaaa