Vault Escape Tunnel & Red River Army Depot Additions
First, I'm dumb and put this on master branch but it's okay.  No one is touching these areas but my dumb ass.

This makes the vaults escape tunnel work out the back of the library, and overhauls the sunset NCR base to be un-NCR'd and made into a radiation death trap robot hellscape.  There is a clear path out the front for Vaulties who try to escape thru it, but hanging out would be inadvisable.

Tested and working.  There might be too much radiation, but I think that more is better.  You shouldn't be hanging around in this area unless you're prepped or scrambling like MAD to get away from it.