Preliminary Tribal Faction Setup
This gets the Tribal faction, the Sulphur Bottom Tribe, into a semi-working state.

The issue with the old codes tribals is that whoever took them out cut bits of code out for some reason, and I'll need help turning them back on properly.  For now tho, these tribals work great for an 'early settler' group to the area.