The Start of Mob Randomization
This PR adds randomized nests to the road just east of Nash, as well as to the Park/Pharmacy block and the residential district just north of that.  These are preliminary and will be expanded across our maps as we move forward.

This particular PR exists specifically to allow Dan to take some of the fixes I got from them on the fly then we can fully move forward more aggressively.