More fixes to Nash
Lots of changes and fixes.
Adds a Graveyard.
Opens the workshop a bit.
Removes the well from the Tribal farm and adds a sink instead.
Adds a shower to the tribal bath house.
Adds some pots and seeds to the bar.
Moves a house onto another house temporarily.
Gives Tribals a lot of lanterns.