Removes broken synth race
Simple change to disable the race Synthetic in the character menu. This DOES NOT affect the Synthetic Lizardperson race, nor does it affect or remove any underlying code for the Synthetics.

## About The Pull Request
The race 'Synthetic' is currently totally broken with invisible sprites, no Species Markings, disabled genitals, etc. This simply removes the broken race as a QOL patchfix until such time in the future as the race's code issues get actually addressed.

## Pre-Merge Checklist
* [ ]  You tested this on a local server.
* [ ]  This code did not runtime during testing.
* [x] You documented all of your changes.

## Changelog 
🆑remove: Disables Synthetic race in Setup Character menu's Species Selection list.
 /🆑