Create [wip] - mappingresources.dmm
adds a .dmm file for mappers to use.
Firstly it has space prefabs and unused buildings and dungeons.
Second it also has a palette with all of the walls, floors, and various other objects because some of them wont appear in the object bar if they arent present on a map that's been opened. To side step this little quirk, I've placed all or at least most of the tiles on this map to call a complete list faster than having to open each and every map file seperately.