Fenny makes pink emotes for people
Sat down with Dan and we mashed out a simple system for flirty emotes.

This includes such bangers as

-*flirt - vague initiation verb

-*flirtaccept - IC agreeing

-*flirtreject - IC rejecting

-*flirtlean - leaning on someone

-*flirtblush - blushing at someone, or something.

-*flirtbrush - brushing someone, or something.

-*flirtcaress - carressing someone or something.

-*flirtcoo - cooing at someone or something.

-*flirteyebrow - quirking an eyebrow at someone or something.

-*flirtgrope - getting a hold of someone or something

-*flirtkiss - kissing someone or something.

-*flirtkissdeep - deeply kissing someone or something.

-*flirtkisslight - lightly kissing someone or something.

-*flirtkissy - making a kissy face at someone or something.

-*flirtlookat - eyeying someone or something up.

-*flirtpinch - pinching someone or something.

-*flirtsmell - getting a smell of someone or something.

-*flirtsniff - a disapproving sniff of someone or something, kinda bratty

-*flirtsnuggle - snuggling someone or something

-*flirtspoonbig - being the big spoon for someone or something

-*flirtspoonlittle - being the little spoon for someone or something

-*flirtwave - languidly waving someone or something over.

-*flirtwink - giving someone or something a wink.

All of these use the paramater system, allowing a player to clarify who and what they may be interacting with when using them using the very simple system of just writing out the emote to a degree.

The benefit of this system is that for players who -want- to throw subtle options to the wind they can make it very clear very rapidly what their intentions appear to be.  Allowing for hook ups to happen more naturally, which should help with the Wallflower problem.

It also includes a *flirtcustom that can use its parameter to write a freeform (love span'd) emote.

Reject and Accept are designed to be used as an IC way to encourage or refuse an advance, and until Dan can get around to automating the system should not be relied on to inform the other person you're stopping them, that is what Local Out of Character is for.

Because sometimes you gotta be a little bratty.

These are ALL range ONE emotes, like subtles are.  Huzzah.