Makes the city a bit more homey
Just needs some changes from Fenny (the billboard thing). I also added a new hospital for the clinic to use instead of the one in Nash. Just outside, I want approval before I push push this though, so.. this will be a draft until Fenny and I talk about it. (That and the old clinic will need replaced.)