AShdown Re-doggification, Small Fixes
Re-adds the forgotten dogs to the Ashdown bar, split by section. Makes the Adult section separated by a curtain with posters indicating its purpose along with a curtain instead of a plain broom-closet style door.

Adds lights to presently lightless house as well as rugs.

Removes the poles from the side booths. (As keeping them would require rotating the entire bar, several rooms, and disconnecting the staff hallways- which are meant to be connected for a reason.)

Also touches up the look of that section of the bar to match the 'feel'.