Berserker Rework!
Makes the Berserker martial arts avaliable for players, with some changes to make it more suitable for PVE and keeping in line with the other martial arts

adds the NO_DRUGS trait to it, as well as changing the choke slam move (which was a knock prone stunlock vs players) into a wrist wrench combo, which disarms the active hand of whatever it's being held as well as causing minor damaga and stamina damage. as well as the AUTO CATCH trait  to make it in line with boar and bass, removes the pierce immunity trait as well as the no dismemberment.