Emergency bug fixes and nerfs
Fixes spriting issues with the new guns
Nerfs the M41 and transfers it to the unique pool to compensate for its stats.
Nerfed Ak-47 and tweaked with lower damage
Adds a small scope to the M3 Magnum to give it a unique flare to itself

Adds the Famas G80 (Fusil d'assault G80), a burst assault rifle.