the penthouse hoehouse
replaces heavens night in the sewer with a large and very full warehouse to pick over (mostly just as filler for right now).
This has been done because Heavens Night now exists on top of the bar in Nash, all the way up, like a penthouse of degeneracy. Actually, not like, thats exactly what it is.