The Faith of Those Not Blinded
Adds the trait blindfaith.  Which is then attached to Shamans and Druids.  It changes their prayer sound from the normal noise to a gong.

Also gives all spirtual prayers the chaplain pray sound, and turns said sound down some.  Because I love your ears and want them to not bleed.