*cuphand
So, this still needs a lil work but the idea is to give you a cupped hand emote to drink water or other 'safe to hold' liquids with.  At the moment its not got a list to stop you from having a handfull of hellwater tho, so maybe we do that first before merging?