Snowflake Loadout+ Biker radio fix
Adds a snowflake loadout for myself that contains the blueprints for a cosmetic variant of the 10mm SMG + A cosmetic variant of the regular brass knuckles.

Also fixes the biker radio because for some reason having the ashdown key first makes it work properly instead of being unable to respond.