Dorime (Porting & Tearing Apart Splurts MERP)
At the time of the first commit I have 62 errors and 115 problems.

They're knot terrible either, mostly that they have a different item slot system as ourselves (which should be able to be shifted over to ours without too much trouble.)

Along with a bunch of missing defines I can't be bothered to grab at this exact moment of posting from their code.

Oh, and duplicate definitions of has_whatevergenitals.

All in all not that far from a test run.