Tries to port tg 59171
This COULD be helping?  I am not even a percent of the coder that these tg guys are, but they say it helps.  so I port it

https://github.com/tgstation/tgstation/pull/59717

It does run, but so much of our really core old code is so out of date as to be comical.  I hate Jay,he butchered Citadels foundational code in his laziness.