Mapping - Joint Shop Changes in front of Nash
Guhh this was a little doozy of a design. I tried to make it a completely, and utterly, infested building full of roaches with different strengths. The roaches on the second level are probably gonna be super easy to kill however. Hoping they wont be. I look forward to seeing how people tackle it in the future, alongside myself with different variations of weaponry.

ALSO I MADE A FEW CHANGES TO SOME OF THE RADROACHES I ADDED WHICH CAN BE PLACED ANYWHERE ELSE ON THE MAP.

I PLAN TO MAYBE MAKE A SUPERMARKET AREA FULL OF RADDY ROACHES ??? MAYBE???

Anyways yes :)