Adjusts the Runechat timers/size and adds Bottom Text :tm:
This increases the size of runechat messages to 200, and lengthens how long they're visible.  Yes, this does mean we could have someone spam them, but its a risk I think is worth taking for QOL.

I am also trying to make it so that it checks for the Trait ERPBOTTOM and places the runechat messages BELOW you, but haven't quite got that working yet.  Could use a real coder

@TalkingCactus
@Superlagg
@OverDriveZ