Starting to port splurts examine/throw/pull icons
So, this like 90% works.  Pretty sure something we're doing is making it go all tits up tho.

What I'd really like to see is some icons attached to the mouse based off intent as well.  This is really a half measure 'fix' that they've created.  It helps, sure, but we can do better.