(MERGE WHEN GOOD) Buffs+ loadout additions
Buffs the needler projectile projectile to have 30 bane instead of 28. No need for loot table adjustments as the difference is extremely minisicle. Adds in improvised needler ammo under the name "jade ammo". Buffs the modification stats of the Type-17, allowing it to be scoped and to have a silencer on it. Adds in new jade projectile icons and capsules.