Ranger Guidebook
While this is a good starting point, it still would do with some good work and maybe one more chapter with the rules of the role.

I fully do not mind the first chapter being editted or requests made on it change wise. It is, afterall, a general tldr foundation for the role.