Halfs the needler lmg rof
Because 800 rounds a minute with an average of +25% from a single mod (1000 rpm) and then maybe, say 33% boost from gold paint is 1330 rpm on a weapon that has high wounding (while most folks wear light armor) and insane bane.

This ACTUALLY brings it down to NORMAL LMG rate of fires.  I'm biting Tox on the b-tooty