Ports Xenoarch from Splurt
Not anything other than the rocktickling part of this but it is working and tested.  Changed out the belt to give you the advanced scanner so that people can actually have fun instead of scratching their rockscrotums for half an hour to get a pile of dog food in a can.