Do something cool with these dungeons.
Dungeons meant for CB. Two are *mostly* finished, though very likely need touchups and finishing touches, as well as loot. The last was meant to be a cave that connected from the forest-- you can actually see where the entrance was meant to be in the forest.

Placing them in templates, because as of right now, that is what they are.