My brain is melting
I have been mapping this for 7 hours straight. It includes;

Small mapping fixes in several locations such as misplaced assets, visual, and fixing some var edits.

It also includes a few new locations and touch ups in the den.

In the tribe the cave of redemption was re-added.

In New Boston the guild breakroom was redone, and a public laundry and bathroom were added, along side several slice of life additions.

If it's broken just scream at me.