(MERGE) Ammo box sprite cull
Culls ammo box sprites to only use 3 types. A full, an empty, and a default rather than 12+ sprites of the same exact sprite. This also fixes the 40mm grenade sprites as well. Stun and other box is untouched as I do not know if we use those in the actual game itself.